Update to .NET 8.0 and fix homing method data type
Upgraded the project to target .NET 8.0 and added a new Release run configuration. Corrected the data type conversion for the homing method in HomingController.

+++ DryveD1API/Controllers/HomingController.cs
@@ -157,7 +157,7 @@ namespace DryveD1API.Controllers
         public void SetMethod(string hostIp, int port, [FromBody] sbyte homingMethod)
         {
             var connection = ModbusSocket.GetConnection(hostIp, port);
-            var data = BitConverter.GetBytes(homingMethod);
+            var data = BitConverter.GetBytes((short)homingMethod);
             var telegram = new Telegram
             {
                 Length = 20
@@ -180,7 +180,7 @@ namespace DryveD1API.Controllers
             try
             {
                 var connection = ModbusSocket.GetConnection(hostIp, port);
-                var data = BitConverter.GetBytes(homingMethod);
+                var data = BitConverter.GetBytes((short)homingMethod);
                 var telegram = new Telegram
                 {
                     Length = 20
